% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PInfDataFunctions.r
\name{allNQQNotNA}
\alias{allNQQNotNA}
\title{Returns input dataframe of short survey questions with 'DitIt' flag.}
\usage{
allNQQNotNA(pDF)
}
\arguments{
\item{pDF}{The dataframe produced by 'getNQTVCs'.}
}
\value{
The long-form dataframe of time varying covariates (returned by
    getNQTVCs) with a column 'DidIt' added, which has the value one if
    any of the questions addressed below (lifetime substance use questions)
    were answered, 0 otherwise.
}
\description{
Used by 'getDididXWave' to create a wide-format dataframe indicating survey
    completion for each of a selected set of waves, by SID.
}
\details{
Used by 'getDidItXWave' to set values for the returned dataframe.
}
\note{
Internal function, NOT EXPORTED
}

