% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PInfDataFunctions.r
\name{checkEligUpdate}
\alias{checkEligUpdate}
\title{Returns TRUE if the Student Eligibility Window (S.E.W.) needs updating.}
\usage{
checkEligUpdate()
}
\value{
A logical variable
}
\description{
The table 'dbo.StudentEligibilityWindow in the database table has 1 row
    for each eligibility period (startdate, enddate)
    for each student who was ever survey-eligible. If Eligibility has
    not expired, the enddate will be set to the last date the update
    sproc (called dbo.StudentEligibilityWindows) was run.
}
\details{
There are no parameters; the sproc simply obtains the date
    that the S.E.W. was last updated, and then the
    most recent date of any survey in the table dbo.SWave (which records
    wave-specific survey data by participating student), and compares
    the two. If the last S.E.W. was more recent
    than the most recent survey, no update is needed (return FALSE);
    otherwise, it is needed (return TRUE).
}
\note{
This function is mainly to be used internally, but a user might
    consider calling it out of curiosity, as to when the last S.E.W.
    table update occurred.
}
