% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PInfDataFunctions.r
\name{createNbhdOnset}
\alias{createNbhdOnset}
\title{Returns data.table of *lagged* binary, up-only 'onset' variables}
\usage{
createNbhdOnset(pInDT, pTHold = 1)
}
\arguments{
\item{pInDT}{A data.table (or dataframe) containing SID plus a lifetime use
(NA, 0, >0) variable for each of c-1 ordered time periods, where c is
the number of columns in the DF.}

\item{pTHold}{The threshold (numerical value) for lifetime use; if the input
use variable value is >= pTHold, then the event 'occurred'.}
}
\value{
A data.table of the same dimension as the input data table, with rows
    transformed into up-only variables suitable for descriptive use.
}
\description{
This stand-alone scoring function takes a wide-form data.table
    or dataframe of 'lifetime use' variables for a (presumably consecutive)
    set of waves, and transforms them into up-only variables. Event onset is
    lagged one period. This data.table can be useful descriptively, because it
    shows the 'network neighborhood' of individuals in the wave before they
    onset to use. It is this neighborhood which presumably influenced onset
    (if there was any such influence).
}
\details{
The logic assigns 1's to the output variables for all waves after the
    first *known* positive instance (value >= pTHold) of the event. It also
    ignores instances where a participant reports the event at wave w, but
    then does not report any lifetime use at some wave w+k. Instead, the
    value of the onset variable will be 1 from the first instance of the event
    on. Also if there are NAs prior to an event being observed, these are all
    output as zeros. These considerations all still apply, even though the
    onset event is lagged one period.
}
\examples{
# Returns a table of 1-period lagged up-only variables which have the value 0
# before the first known lifetime alcohol use (if any such times are
# available), and 1 thereafter. The pTHold value of 1 means *any* use.
alcLifeLag <- makeTVTbl(getNQTVCs(c(1, 3, 4), c(3, 4, 5, 6, 30)))
alcOnsetLag <- createOnset(alcLifeLag, pTHold = 1)
}

