% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PInfDataFunctions.r
\name{getDiditXWave}
\alias{getDiditXWave}
\title{Returns dataframe with flags indicating survey completion}
\usage{
getDiditXWave(pWavVec, pSchVec)
}
\arguments{
\item{pWavVec}{An ordered numeric vector of wave SIDs to be included in the
analysis.
Waves 2, 6, and 10 are summer waves; no Short Survey or Long Survey data
were collected. Hence these waves should not ever be specified.}

\item{pSchVec}{A numeric vectorof School IDs to be included in the analysis.}

\item{pElig}{The minimum number of eligible waves for SID to be included.}

\item{pDid}{The minimum number of survey-completed waves for SID to be
included.}
}
\value{
A (#eligible SIDs x length(pWavVec)+1) dataframe. Col 1 is SID,
    Cols 2 on (each named 'wK', where K is the kth ordered wave in pWavVec)
    are flags with the integer 1 if individual SID 'completed' a survey
    for that wave (completion is determined in the function 'allNQQNotNA')
}
\description{
Returns dataframe with flags indicating survey completion
}
\examples{
# gets dataframe including SIDs for waves 1, 3, 4, and 5, schools 3-6 and
# 30 (one school district); each included SID was eligible for the Short
# survey at least once, and completed at least 1 survey in these waves.
sidVec <- getSIDSet(c(1, 3, 4, 5), c(3, 4, 5, 6, 30), 1, 1)
}
