% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PInfDataFunctions.r
\name{getDOB}
\alias{getDOB}
\title{Returns a data.table with DOBs for all SIDs input}
\usage{
getDOB(pSIDVec)
}
\arguments{
\item{pSIDVec}{A vector of SIDs for which DOB is wanted. These will
normally be part of an 'Analysis Set', i.e. generated by
'getNetworkSet' or 'getSIDSet'.}
}
\value{
A data.table with two columns: SID and DOB. SID is integer,
    DOB is POSIXct/POSIxt
}
\description{
This function can be used to calculate an age variable
}
\details{
Makes a call to the database using RODBC
}
\examples{
# Gets DOB for a couple of specific SIDs
aFewDOBs <- getDOB(c(10, 11, 12))
}

