% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PInfDataFunctions.r
\name{getEligNodes}
\alias{getEligNodes}
\title{Returns a vector of SIDs who were Short Survey eligible on any input wave}
\usage{
getEligNodes(pWavVec, pSchVec)
}
\arguments{
\item{pWavVec}{An ordered numeric vector of wave SIDs to be included in the
analysis.
Waves 2, 6, and 10 are summer waves; no Short Survey or Long Survey data
were collected. Hence these waves should not ever be specified.}

\item{pSchVec}{A numeric vectorof School IDs to be included in the analysis.}
}
\value{
A (#eligible SIDs)-length numerical vector of SIDs.
}
\description{
Returns a vector of SIDs who were Short Survey eligible on any input wave
}
\details{
Used by other functions in this package (e.g. getNetworkSet,
    makeTVTbl, )
}
\note{
This function just repeatedly calls 'getSWNodes' for each requested
    school & wave combination
}
\examples{
# gets vector of SIDs for waves 1, 3, 4, and 5, schools 3-6 and 30 (one
# school district)
sidVec <- getSIDSet(c(1, 3, 4, 5), c(3, 4, 5, 6, 30), 1, 1)
}
