% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PInfDataFunctions.r
\name{getFixedCovs}
\alias{getFixedCovs}
\title{Returns a dataframe of fixed covariates for a set of input schools}
\usage{
getFixedCovs(pSchVec)
}
\arguments{
\item{pSchVec}{A vector of (integer) School IDs; any existing fixed
covariate rows in the database where the participant is shown
as currently in one of these schools will be included.}
}
\value{
A dataframe that has the same format as the database table
    'PInf.dbo.SMaster'. Column names fairly accurately describe
    column contents, but one should consult the qq codebook for
    precise definition.
}
\description{
This is generally used to pull this data from the database, and
    subsequently subset it for a particular Analysis Set.
}
\details{
If a participant switches schools (which happens regularly),
    this function will only pull his/her data if his/her *current*
    school is in the input vector.
}
\examples{
# Gets table of fixed covariats for all the schools in 1 district
fc <- getFixedCovs(c(3,4,5,6,30)
}
