% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PInfDataFunctions.r
\name{getMSRIDVecs}
\alias{getMSRIDVecs}
\title{Creates a list of *Row* ID vectors for each middle school requested}
\usage{
getMSRIDVecs(mSchVec, sidRowID)
}
\arguments{
\item{mSchVec}{A numeric vector of middle school IDs to be separated
at Wave 1 (only, because in this study, participants are in middle
school only for Wave 1) by structural zeros.}

\item{sidRowID}{A dataframe-format crosswalk of the SIDs and corresponding
Row IDs in the analysis set being developed by the function that
calls this one.}
}
\value{
A length(mSchVec) list of Row IDs of individuals within
    each of the middle schools with IDs in 'mSchVec'.
}
\description{
The function is used for multiple middle school districts to
    add structural zeros to a wave 1 network edge table.
}
\details{
Adding structural zeros is an option of 'getNetworkSet', which
    is where this function is currently used. In the PInf study, this is
    only meaningful for middle schools from the same school district at
    Wave 1.
}
\examples{
# Gets dataframe including SIDs for the wave vector 'mSchVec' and the
# two-column dataframe 'sidRowID'.
msIDList <- getMSRIDVecs (mSchVec, sidRowID)
}

