% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PInfDataFunctions.r
\name{getNameAsString}
\alias{getNameAsString}
\title{Returns the name of a variable as a string.}
\usage{
getNameAsString(pVarName)
}
\arguments{
\item{pVarName}{A variable name (no quotes or anything)}
}
\value{
The variable name as a string
}
\description{
This function just removes the need to look up how to do this (fairly
    common) operation, which I seem to have to do all the time, because
    I can't remember the (little used) syntax.
}
\examples{
# Returns c("myVariable")
vStr <- getNameAsString(myVariable)
}

