% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PInfDataFunctions.r
\name{getNetwork}
\alias{getNetwork}
\title{Returns a dataframe of network edges for input schools and waves}
\usage{
getNetwork(pWavVec, pSchVec)
}
\arguments{
\item{pWavVec}{An ordered numeric vector of wave SIDs to be included in the
analysis.
Waves 2, 6, and 10 are summer waves; no Short Survey or Long Survey data
were collected. Hence these waves should not ever be specified.}

\item{pSchVec}{A vector of (integer) School IDs.}
}
\value{
A dataframe that has the same format as the database table
    'PInf.dbo.SAffiliation'. Column names fairly accurately describe
    column contents, but one should consult the qq codebook for
    precise definition.
}
\description{
This is generally used to pull this data from the database, and
    subsequently subset it for a particular Analysis Set. It is
    currently used by 'getNetworkSet'.
}
\details{
Chooser was asked for more information on an individual (if any)
    designated as Chooser's 'very best friend'. These variables are
    missing for any alters not so designated, or if Chooser did not
    make this designation at all.
}
\examples{
# Gets edgelist (with addtional info on some alters) for individuals in
# waves 1, 3, and 4, schools 3:6 and 30 (one school district).
net <- getNetwork(c(1,3,4), c(3,4,5,6,30))
}
