% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PInfDataFunctions.r
\name{getNQTVCs}
\alias{getNQTVCs}
\title{Returns a data.table of Short-Survey time-varying covariates}
\usage{
getNQTVCs(pWavVec, pSchVec)
}
\arguments{
\item{pWavVec}{An ordered numeric vector of wave SIDs to be included in the
analysis.
Waves 2, 6, and 10 are summer waves; no Short Survey or Long Survey data
were collected. Hence these waves should not ever be specified.}

\item{pSchVec}{A vector of (integer) School IDs.}
}
\value{
A dataframe that is a subset of the database table
    'PInf.dbo.SWave'., containing only the items asked on the 'Short Survey'.
    Column names fairly accurately describe
    column contents, but one should consult the qq codebook for
    precise definition.
}
\description{
This is generally used to pull this data from the database, and
    subsequently subset it for a particular Analysis Set. It is
    currently used by 'makeTVTbl'.
}
\examples{
# Gets a long-form data.table (each row represents an individual and a wave)
# of survey data from the 'Short Survey', for
# waves 1, 3, and 4, schools 3:6 and 30 (one school district).
tvcs <- getNQTVCs(c(1, 3, 4), c(3, 4, 5, 6, 30))
}
