% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PInfDataFunctions.r
\name{getSIDSet}
\alias{getSIDSet}
\title{Creates a vector of SIDs that are consistent with some Analysis Set}
\usage{
getSIDSet(pWavVec, pSchVec, pElig = 1, pDid = 1)
}
\arguments{
\item{pWavVec}{An ordered numeric vector of wave SIDs to be included in the
analysis.
Waves 2, 6, and 10 are summer waves; no Short Survey or Long Survey data
were collected. Hence these waves should not ever be specified.}

\item{pSchVec}{A numeric vectorof School IDs to be included in the analysis.}

\item{pElig}{the minimum number of eligible waves for SID to be included}

\item{pDid}{the minimum number of completed waves for SID to be included}
}
\value{
An Analysis Set-length numerical vector of SIDs.
}
\description{
This function is a bit slow, probably because it uses calls to
    'getEligXWave' and 'getDiditXWave'...but it still runs in a couple
    of minutes even for fairly large Analysis Sets, and this design
    has the advantage of not requiring the output from these functions
    to be created ahead of time and passed in.
}
\details{
Used by other functions in this package (e.g. getNetworkSet).
    Could be used stand-alone but is not really designed to be.
}
\examples{
# Obtains the SID set for a network analysis:
# Use the last element of 'netList', which is a vector of the SIDs in
# the analysis
}

