% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PInfDataFunctions.r
\name{getSWNodes}
\alias{getSWNodes}
\title{Returns a vector of SIDs for Short or Long-form survey-eligibles}
\usage{
getSWNodes(pWav, pSch, pNQW = 1, pFQW = 0)
}
\arguments{
\item{pWav}{A specific survey wave.
Waves 2, 6, and 10 are summer waves; no Short Survey or Long Survey data
were collected. Hence these waves should not ever be specified.}

\item{pSch}{A specific School ID.}

\item{pNQW}{If 1, short-survey-eligible SIDs in this wave and school
are returned (this would also include long-survey-eligibles as a
subset).}

\item{pFQ}{If 1, only long-survey-eligible SIDs in this wave and school are
returned.}
}
\value{
An integer vector of SIDs
}
\description{
Output from this function is useful in creating subsets of data meeting
    survey-eligibility requirements.
}
\examples{
# Returns a vector of SIDs of individuals who were short-survey (NQ)
# eligible at wave 1, from school #3.
nodes <- getNQTVCs(pWav = 1, pSch = 3, pNQW = 1)
}

