% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PInfDataFunctions.r
\name{getTVCCols}
\alias{getTVCCols}
\title{Returns a vector of column (item) names for a scale or variable}
\usage{
getTVCCols(pVar)
}
\arguments{
\item{pVar}{A string giving one of the abbreviations for a variable
understood by the function. See below for a list of possible values.}
}
\value{
A vector of strings corresponding to the names of the items
    comprising a variable in the database table 'PInf1.dbo.SWave'
}
\description{
Used by 'getNQTVCs' to obtain the items necessary for a scale (or in some
    cases, just a single items, e.g. 'number of days drank, last 30'). It
    has the effect of generalizing 'getNQTVCs' so that the call to that
    function is compact, and does not require the user to go find the
    names of the items in any given scale.
}
\note{
Choices for pVar are:
    AB (Antisocial behavior),
    OV (Others victimization of you),
    YV (You victimizing others),
    T3 (Tobacco, freq of use in last 30 days),
    E3 (E-tobacco, freq of use in last 30 days),
    C3 (Chewing Tobacco, freq of use in last 30 days),
    A3 (Alcohol, freq of use in last 30 days),
    B3 (Binge drinking, freq in last 30 days),
    M3 (Marijuana use, freq in last 30 days),
    AL (Alcohol, freq of use, lifetime)
}
\examples{
# Internal function, not normally available to users. See 'makeTVTbl'
# source code to see how it is used.
}

