% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PInfDataFunctions.r
\name{\%w/o\%}
\alias{\%w/o\%}
\title{'Without' operator (complement of %in%):}
\usage{
x \%w/o\% y
}
\description{
'Without' operator (complement of %in%):
}
\details{
A binary operator, i.e. x %w/o% y (x and y vectors, or lists) gives
   the elements of x that are NOT IN y.
}
