% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PInfDataFunctions.r
\name{make01MemberMatrix}
\alias{make01MemberMatrix}
\title{Creates a matrix indicating membership in a group based on a variable}
\usage{
make01MemberMatrix(pInDT, pPrefix)
}
\arguments{
\item{pInDT}{a data.table with two columns: SID and some integer
variable}

\item{pPrefix}{a string indicating the prefix to give to each
category-specific column of the output}
}
\value{
A data.table with the same two columns as the input DT, plus
   k new columns, one for each unique category (i.e. value) in the
   2nd column of pInDT. Let 'str' be the value of 'pPrefix'; and suppose
   the categories are numbered 1, 2, ... then
   the columns will have the names str1, str2, etc.
}
\description{
Membership is assumed to be described by a set of integers stored in
the variable.
}
