% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PInfDataFunctions.r
\name{makeCCVec}
\alias{makeCCVec}
\title{Returns a list of vectors representing RSiena composition change.}
\usage{
makeCCVec(pElig)
}
\arguments{
\item{pElig}{An n x 1+w dataframe, with 1 row for every SID who was
*survey-eligible* for any of the waves of interest, and 1 col for
each wave, containing 1 if the SID was survey-eligible that wave,
and 0 if not (col 1 is SID).}
}
\value{
A list with n elements (one for each row of the inut DF). Each
    element is a vector of pairs of arrival and departure times from
    the network, as explained in the RSiena manual under 'method of
    changing composition'.
}
\description{
The output of this function may be input directly to the RSiena function
    'sienaCompositionChange'.
}
\details{
The input dataframe must be the object returned by
    'getEligXWave', or the equivalent.
}
\examples{
# Returns a list of composition change vectors for the SIDs
# in waves 1, 3, and 4, and schools 3, 4, 5, 6, and 30.
ccDF <- makeCCVec(getEligXWave(pWavVec = c(1,3,4),
    pSchVec = c(3,4,5,6,30)))
}
