% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PInfDataFunctions.r
\name{makeTVTbl}
\alias{makeTVTbl}
\title{Returns a data.table of a time varying behavior variable}
\usage{
makeTVTbl(pTVTbl, pSIDs, pVar = "X", pCut = c(0), pMaxNA = 1)
}
\arguments{
\item{pSIDs}{A numeric vector of SIDs of the individuals in this Analysis
Set (i.e. for example, created by 'getNetworkSet'). Ensures the table
created is consistent with such a set.}

\item{pVar}{A string giving the name of the variable to create. These are
defined in 'getNQTVCs'. Options are listed below.}

\item{pCut}{A numeric vector of cut points for classifying values of
the requested variable (see Base R function 'cut'). Intervals are
closed on the left, open on the right.}

\item{pMaxNA}{The maximum number of NAs among the variable's items
before the returned value for the particular subject-wave is set to NA}

\item{pCCTbl}{A dataframe or data.table of all the time-varying covariates
in long-form (1 row per subject per wave) produced by the function
   'getNQTVCs'. It is an image of the database table PInf1.dbo.SWave.}
}
\value{
The output table format is w+1 columns (w=#waves present in
    pCCTbl); first column is an ordered set of SIDs, columns 2:(w+1)
    are the scored (mean of items) value of the requested behaivor
    variable for the w waves.
}
\description{
Returns a data.table of a time varying behavior variable
}
\details{
May be used to create tables that can be easily converted to SAOM
    dependent behavior or time-varying covariate (predictor) variables by
    just selecting the last w columns (see example below). Generally such
    variables will be consistent with a network Analysis Set (of
    individuals) produced by 'getNetworkSet', in which case 'pSIDs' can
    simply be taken from the last list element of the output of the
    latter function. Of course, if you wanted to create variables for any
    longitudinal analysis (i.e. score them, select the individuals and
    waves you wanted), you could use this function as well, but you would
    need to generate the pSIDs by directly executing 'getSIDSet'.
}
\note{
Choices for pVar are defined in 'getTVCCols'. They are:
    AB (Antisocial behavior),
    OV (Others victimization of you),
    YV (You victimizing others),
    T3 (Tobacco, freq of use in last 30 days),
    E3 (E-tobacco, freq of use in last 30 days),
    C3 (Chewing Tobacco, freq of use in last 30 days),
    A3 (Alcohol, freq of use in last 30 days),
    B3 (Binge drinking, freq in last 30 days),
    M3 (Marijuana use, freq in last 30 days),
    AL (Alcohol, freq of use, lifetime up to now).
}
\examples{
# When the network part of an analysis set is created by 'getNetworkSet',
# the last element of the output is a vector of participating SIDs. We
# use that as part of the input to the present function.
netList <- getNetworkSet(pWavVec = c(1,3,4), pSchVec = c(3,4,5,6,30),
    pElig = 2, pDid = 2, pTyp = "BF", pOut = "SP", pS0 = "S0")
# Create the time varying behavior table
ccTbl <- getNQTVCs(pWavVec = c(1,3,4), pSchVec = c(3,4,5,6,30))
# Creates a data.table of antisocial behavior variables x subject & wave,
# for the 3 waves & other criteria used above, grouped into 4 categories,
# and with no more than 3 NAs out of the 6 items.
abTbl <- makeTVTbl(ccTbl, netList[[4]], pVar = "AB",
    pCut = c(0, .5, 1, 5, 20), pMaxNA = 3)
}

