% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PInfDataFunctions.r
\name{s0Assemble}
\alias{s0Assemble}
\title{Returns a data.table with structural 0's for w1 middle schools}
\usage{
s0Assemble(pMSchVec)
}
\arguments{
\item{pMSchVec}{A list of vectors, each containing the Wave 1 Analysis Set
RowIDs (because Network Sets use RowIDs, as required by RSiena and
network) for a specific middle school.}
}
\value{
A data.table in edgelist format; cols are RowID (chooser),
    RowID (chosen), bff (value of the relationship, an integer >= 0),
    and WID (wave ID #)
}
\description{
This is a 'helper' function which, along with s0Make, adds rows of
    structural zeros for individuals from different middle schools
    who, however, will be going to the same high school the next
    year...and are thus considered part of the same network, for
    analysis purposes.
}
\note{
This function should not normally be used stand-alone. To see how it
    is used in context, consult the source code for 'getNetworkSet'.
    NOT EXPORTED
}

