% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PInfDataFunctions.r
\name{s0Make}
\alias{s0Make}
\title{Returns a data.table of structural 0 edges between 2 schools}
\usage{
s0Make(pVec1, pVec2)
}
\arguments{
\item{pVec1}{A vector of the RowIDs (note: not SIDs!) in middle school 1.}

\item{pVec2}{A vector of the RowIDs in middle school 2.}
}
\value{
A data.table in edgelist format; cols are RowID (chooser),
    RowID (chosen), bff (value of the relationship, an integer >= 0),
    and WID (wave ID #).
}
\description{
This is a 'helper' function which, along with s0Assemble, adds rows of
    structural zeros for individuals from different middle schools
    who, however, will be going to the same high school the next
    year...and are thus considered part of the same network, for
    analysis purposes.
}
\details{
To make connections between two wave 1 middle schools impossible,
    all you need to do is create an edge with value '10 (structural zero,
    in RSiena terminology) between each pair of individuals in the two
    schools. This function accomplishes that in a simple but efficient way.
    The function 's0Assemble' takes all the data.table objects
    created here and assembles them into one big table to return to the
    calling function.
}
\note{
This function should not normally be used stand-alone. To see how it
    is used in context, consult the source code for 'getNetworkSet'.
    NOT EXPORTED
}
