% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PInfDataFunctions.r
\name{schWvExists}
\alias{schWvExists}
\title{Returns logical TRUE if requested wave includes requested school}
\usage{
schWvExists(pWav, pSch)
}
\arguments{
\item{pWav}{The wave number of interest. Cannot be 2, 6, or 10.}

\item{pSch}{The school ID number of interest.}
}
\value{
Logical variable; TRUE if the requested school and wave has any
    observations, otherwise FALSE.
}
\description{
This information is inherent in the study design; middle schools are only
    observed in Wave 1 (and high schools are not), whereas high schools
    are only observed in waves 3:5, 7:9, and 11.
}
\note{
: Used internally only-- NOT EXPORTED
}
