% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PInfDataFunctions.r
\name{allNQQNotNA}
\alias{allNQQNotNA}
\title{Returns input dataframe of short survey questions with 'DitIt' flag.}
\usage{
allNQQNotNA(pDF)
}
\arguments{
\item{pDF}{The dataframe produced by 'getNQTVCs'.}
}
\value{
The long-form dataframe of time varying covariates (returned by
    getNQTVCs) with a column 'DidIt' added, which has the value one if
    any of the questions addressed below (lifetime substance use questions)
    were answered, 0 otherwise.
}
\description{
Used by 'getDididXWave' to create a wide-format dataframe indicating survey
    completion for each of a selected set of waves, by SID.
}
\details{
Used by 'getDidItXWave' to set values for the returned dataframe.
}
\note{
Internal function, NOT EXPORTED
}
\examples{
# Returns the table of time varying covariates produced by 'getNQTVCs' with
# the column 'DidIt' added, for the schools and waves in pSchVec and pWavVec.
tvc <- allNQQNotNA(getNQTVCs(pWavVec = c(1, 3, 4),
                             pSchVec = c(3, 4, 5, 6, 30)))
}

